- Implemented Dugan's scattering model
- Added forward/diffuse/backward/transparent/absorption modes for materials
- Added "teleport to previous"
- Referring to materials by name
- Added SUMFLUX, etc. to Synrad's formula editor
- Few long->size_t conversions

+++ Region_mathonly.h
@@ -47,7 +47,7 @@ public:
 
 	//Loaded from files
 	Vector nbDistr_MAG; //MAG file points
-	int nbDistr_BXY; //BXY file points
+	size_t nbDistr_BXY; //BXY file points
 	Distribution2D *Bx_distr,*By_distr,*Bz_distr; //pointer to the B field distribution (if file-based)
 	Distribution2D *beta_x_distr,*beta_y_distr,*eta_distr,*etaprime_distr,*e_spread_distr; //pointer to the beta field distribution (if file-based)
 	Vector Bx_dir,By_dir,Bz_dir; //direction in which MAG files are oriented (in their second line)
@@ -56,7 +56,7 @@ public:
 
 	//Calculated data
 	std::vector<Trajectory_Point> Points;
-	int nbPointsToCopy; //passes info about the number of points that needs to be read from the buffer on loading
+	size_t nbPointsToCopy; //passes info about the number of points that needs to be read from the buffer on loading
 	Vector startPoint,startDir,B_const,limits;//AABBmin,AABBmax
 	Quadrupole quad;
 	//int selectedPoint;
